# docs/swagger.yaml
definitions:
  middle.ResCode:
    enum:
    - 200
    - 201
    - 202
    - 203
    - 204
    - 205
    - 206
    - 207
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeInvalidParam
    - CodeUserExist
    - CodeUserNotExist
    - CodeInvalidPassword
    - CodeServerBusy
    - CodeNeedLogin
    - CodeInvalidToken
  middle.ResponseData:
    properties:
      code:
        $ref: '#/definitions/middle.ResCode'
      data: {}
      msg: {}
    type: object
...
  model.CasbinInReceive:
    properties:
      authorityId:
        description: 权限id
        type: integer
    type: object
  model.CasbinInfo:
    properties:
      method:
        description: 方法
        type: string
      path:
        description: 路径
        type: string
    type: object
...
  model.UpdateCasbinInput:
    properties:
      authorityId:
        description: 权限id
        type: integer
      casbinInfos:
        items:
          $ref: '#/definitions/model.CasbinInfo'
        type: array
    type: object
...
info:
  contact: {}
paths:
  /api/authority/getPolicyPathByAuthorityId:
    get:
      consumes:
      - application/json
      parameters:
      - description: 权限id, 权限模型列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CasbinInReceive'
      produces:
      - application/json
      responses:
        "200":
          description: 获取权限列表,返回包括casbin详情列表
          schema:
            allOf:
            - $ref: '#/definitions/middle.ResponseData'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取权限列表
      tags:
      - Casbin
  /api/authority/updateCasbinByAuthority:
    post:
      consumes:
      - application/json
      parameters:
      - description: 权限id, 权限模型列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.UpdateCasbinInput'
      produces:
      - application/json
      responses:
        "200":
          description: 通过角色更新接口权限
          schema:
            allOf:
            - $ref: '#/definitions/middle.ResponseData'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 通过角色更新接口权限
      tags:
      - Casbin
...
  /api/k8s/configmap/del:
    delete:
      consumes:
      - application/json
      description: 删除Configmap
      operationId: /api/k8s/configmap/del
      parameters:
      - description: Configmap名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "删除成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 删除Configmap
      tags:
      - Configmap
  /api/k8s/configmap/detail:
    get:
      consumes:
      - application/json
      description: 获取Configmap详情
      operationId: /api/k8s/configmap/detail
      parameters:
      - description: Configmap名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data":v1.Deployment }'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 获取Configmap详情
      tags:
      - Configmap
  /api/k8s/configmap/list:
    get:
      consumes:
      - application/json
      description: 查看Configmap列表
      operationId: /api/k8s/configmap/list
      parameters:
      - description: 过滤
        in: query
        name: filter_name
        type: string
      - description: 命名空间
        in: query
        name: namespace
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 分页限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": }'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 查看Configmap列表
      tags:
      - Configmap
  /api/k8s/configmap/update:
    put:
      consumes:
      - application/json
      description: 更新Configmap
      operationId: /api/k8s/configmap/update
      parameters:
      - description: 无状态控制器名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: 更新内容
        in: query
        name: content
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "更新成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 更新Configmap
      tags:
      - Configmap
  /api/k8s/daemonset/del:
    delete:
      consumes:
      - application/json
      description: 删除DaemonSet
      operationId: /api/k8s/DaemonSet/del
      parameters:
      - description: DaemonSet名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "删除成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 删除DaemonSet
      tags:
      - DaemonSet
  /api/k8s/daemonset/detail:
    get:
      consumes:
      - application/json
      description: 获取DaemonSet详情
      operationId: /api/k8s/DaemonSet/detail
      parameters:
      - description: DaemonSet名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data":v1.Deployment }'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 获取DaemonSet详情
      tags:
      - DaemonSet
  /api/k8s/daemonset/list:
    get:
      consumes:
      - application/json
      description: 查看DaemonSet列表
      operationId: /api/k8s/DaemonSet/list
      parameters:
      - description: 过滤
        in: query
        name: filter_name
        type: string
      - description: 命名空间
        in: query
        name: namespace
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 分页限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": }'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 查看DaemonSet列表
      tags:
      - DaemonSet
  /api/k8s/daemonset/update:
    put:
      consumes:
      - application/json
      description: 更新DaemonSet
      operationId: /api/k8s/DaemonSet/update
      parameters:
      - description: 无状态控制器名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: 更新内容
        in: query
        name: content
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "更新成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 更新DaemonSet
      tags:
      - DaemonSet
...
swagger: "2.0"